Add market update throttling for perpetuals
Introduces a mechanism to throttle market updates in Perpetuals by storing the last update date in Preferences and only updating markets if more than 7 days have passed. Adds a Date extension for checking if a date is outdated by a given interval or number of days. Refactors related view and view model logic to support this change.