Update toolbar dismiss actions and price alert logic (#1192)
* Update toolbar dismiss actions and price alert logic

Replaces 'cancel' with 'done' for toolbar dismiss items in TextMessageScene and WalletConnectorNavigationStack for consistency. Also updates AssetSceneViewModel to show price alerts based solely on the presence of alerts, not the enabled state.

* Change toolbar dismiss button from Done to Cancel

Updated the toolbar dismiss item title from 'Done' to 'Cancel' in WalletConnectorNavigationStack for improved clarity in user navigation.