Remove chatClosed handler and add script to hide close button (#1526)
* Remove chatClosed handler and add script to hide close button

Eliminates the unused chatClosed handler from ChatwootHandler and related code in ChatwootWebViewModel. Adds a WKUserScript to hide the chat close button in the web view for improved UI control.

* Refactor injected CSS script in ChatwootWebViewModel

Simplified the JavaScript code for injecting CSS by removing the IIFE and using 'textContent' instead of 'innerHTML' for the style element. This improves readability and aligns with best practices for DOM manipulation.